cambie tipos de datos a double pero aun no se actualiaz el 2%+

+++ src/Character.cpp
@@ -1,12 +1,12 @@
 // Character.cpp
 #include "Character.h"
 
-Character::Character(const std::string& name, int hp, int atk, int def, int lck, int spd)
+Character::Character(const std::string& name, double  hp, double  atk, double  def, double  lck, double  spd)
     : name(name), hp(hp), atk(atk), def(def), lck(lck), spd(spd) {}
 
 Character::~Character() {}
 
-void Character::takeDamage(int amount) {
+void Character::takeDamage(double  amount) {
     hp -= amount;
     if (hp < 0) hp = 0;
 }
@@ -19,23 +19,23 @@ const std::string& Character::getName() const {
     return name;
 }
 
-int Character::getHP() const {
+double  Character::getHP() const {
     return hp;
 }
 
-int Character::getDefense() const {
+double  Character::getDefense() const {
     return def;
 }
 
-int Character::getAttack() const {
+double  Character::getAttack() const {
     return atk;
 }
 
-int Character::getLuck() const {
+double  Character::getLuck() const {
     return lck;
 }
 
-int Character::getSpeed() const {
+double  Character::getSpeed() const {
     return spd;
 }
