+++ src/Enemy.h
@@ -7,7 +7,7 @@
 
 class Enemy : public Character {
 public:
-    Enemy(const std::string& type, int hp, int atk, int def, int lck, int spd);
+    Enemy(const std::string& type, double  hp, double  atk, double  def, double  lck, double  spd);
     void attack(Character& target) override;
 };
